requirements.txt
flask-restful==0.3.9
Flask==2.0.3
pyjwt
python-dotenv==0.19.2
djangorestframework==3.13.1
redis==2.10.5
gunicorn==20.0.4
Django==2.2.26
yt-dlp==2022.3.8.2
django-crontab==0.7.1
django-cors-headers==3.7.0
pymemcache==3.5.1
requests